Fixing buienradar data URI
It seems that the URI to access XML data from the Buienradar API was changed to https://data.buienradar.nl/1.0/feed/xml. The old uri no longer returns a result in NodeRED, quiet possibly because this starts a redirect and the NodeJS request might not be so keen on allowing that. Changing the URI to the new format fixed the problem.